+++ frontend-api/src/main/kotlin/no/nav/melosys/tjenester/gui/dto/trygdeavgift/TrygdeavgiftsperiodeDto.kt
@@ -19,7 +19,7 @@ data class TrygdeavgiftsperiodeDto(
             trygdeavgiftsperiode.periodeTil,
             trygdeavgiftsperiode.grunnlagMedlemskapsperiode.trygdedekning,
             trygdeavgiftsperiode.grunnlagInntekstperiode.type,
-            trygdeavgiftsperiode.trygdesats,
+            trygdeavgiftsperiode.trygdesats.toDouble(),
             trygdeavgiftsperiode.trygdeavgiftsbeløpMd.verdi.toInt()
         )
 }
