+++ integrasjon/src/main/kotlin/no/nav/melosys/integrasjon/trygdeavgift/dto/TrygdeavgiftsperiodeDto.kt
@@ -1,7 +1,9 @@
 package no.nav.melosys.integrasjon.trygdeavgift.dto
 
+import java.math.BigDecimal
+
 data class TrygdeavgiftsperiodeDto(
     val periode: DatoPeriodeDto,
-    val sats: Double,
+    val sats: BigDecimal,
     val månedsavgift: PengerDto,
 )
